+++ clarity/tests/sip10-reserve_test.ts
@@ -4,7 +4,7 @@ import {
   Clarinet,
   Tx,
   types,
-} from "https://deno.land/x/clarinet@v0.5.2/index.ts";
+} from "https://deno.land/x/clarinet@v0.6.0/index.ts";
 import { assert } from "https://deno.land/std@0.90.0/testing/asserts.ts";
 
 Clarinet.test({